% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cmu.textstat-package}
\alias{cmu.textstat}
\alias{cmu.textstat-package}
\title{CMU Text Analysis datasets and utility package}
\description{
Datasets and utility code for the text analysis course.
}
\author{
\strong{Maintainer}: David Brown \email{dwb2@andrew.cmu.edu} (\href{https://orcid.org/0000-0001-7745-6354}{ORCID})

Authors:
\itemize{
  \item Alex Reinhart \email{areinhar@stat.cmu.edu} (\href{https://orcid.org/0000-0002-6658-514X}{ORCID})
}

}
\keyword{internal}
