% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalist.R
\docType{data}
\name{federalist_meta}
\alias{federalist_meta}
\title{Federalist paper metadata}
\format{
Data frame with four columns:
\describe{
\item{doc_id}{Document ID, giving the number of the Federalist paper}
\item{author_id}{Author: Hamilton, Madison, Jay, \code{Hamilton_Madison} (collaboration of both), or Disputed}
\item{title}{Original title}
\item{date}{Original date, if available}
}
}
\usage{
federalist_meta
}
\description{
Metadata for each Federalist Paper, including the author, title, and date (if known).
}
\seealso{
\link{federalist_papers}
}
\keyword{datasets}
