% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalist.R
\docType{data}
\name{federalist_papers}
\alias{federalist_papers}
\title{Federalist papers}
\format{
readtext data frame with two columns:
\describe{
\item{doc_id}{Document ID, giving the number of the Federalist paper}
\item{text}{Full text of the paper}
}
}
\usage{
federalist_papers
}
\description{
Contains the full text of the 85 Federalist Papers.
}
\seealso{
\link{federalist_meta}, \code{\link[readtext:readtext]{readtext::readtext()}}
}
\keyword{datasets}
