% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micusp.R
\docType{data}
\name{micusp_mini}
\alias{micusp_mini}
\title{Mini excerpt of MICUSP}
\format{
\describe{
\item{doc_id}{Document ID (from MICUSP). First three characters give the document category.}
\item{text}{Full text of the student paper}
}
}
\usage{
micusp_mini
}
\description{
170 documents from the Michigan Corpus of Upper-Level Student Papers.
}
\seealso{
\link{micusp_meta}
}
\keyword{datasets}
