% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenplays.R
\docType{data}
\name{screenplays}
\alias{screenplays}
\title{50 tagged screenplays}
\format{
\describe{
\item{doc_id}{Document ID (filename of original screenplay)}
\item{text}{Full tagged text of screenplay}
}
}
\usage{
screenplays
}
\description{
Screenplays from 50 movies. Tagged HTML-style with \verb{<DIRECTION>} and
\verb{<DIALOGUE>} tags, to distinguish stage directions from dialogue. Other tags
indicate the characters speaking.
}
\keyword{datasets}
