% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\docType{data}
\name{sentiment_data}
\alias{sentiment_data}
\title{Novels for sentiment analysis}
\format{
readtext data frame with 4 observations
}
\usage{
sentiment_data
}
\description{
Contains the full text of \emph{Madame Bovary}, \emph{A Portrait of the Artist as a
Young Man}, \emph{Ragged Dick}, and \emph{The Rise of Silas Lapham}.
}
\keyword{datasets}
