% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shakespeare.R
\docType{data}
\name{shakespeare_corpus}
\alias{shakespeare_corpus}
\title{Corpus of Shakespeare's plays}
\format{
readtext data frame with 37 plays, each with \code{doc_id} (giving type
and title) and \code{text}.
}
\usage{
shakespeare_corpus
}
\description{
Full tagged text of Shakespearean plays, in XML form with stage direction,
scenes, and dialogue tagged.
}
\seealso{
\code{\link[=from_play]{from_play()}} to extract subsets of text from each play.
}
\keyword{datasets}
