% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_samples.R
\docType{data}
\name{us_sample}
\alias{us_sample}
\alias{uk_sample}
\title{Small subset samples from BAWE and MICUSP corpora}
\format{
\code{readtext} objects each containing 100 observations with document ID
and fully part-of-speech tagged text.

An object of class \code{readtext} (inherits from \code{data.frame}) with 100 rows and 2 columns.
}
\usage{
us_sample

uk_sample
}
\description{
100 observations each from BAWE (student writing from the UK) and MICUSP
(student writing from the US) corpora.
}
\keyword{datasets}
